DebugSummitFromFZQ02
Detailed in UPDATE.md.

+++ DDLSurvivor/PacmanFight/BackEnd/pacman_object.h
@@ -8,7 +8,7 @@ class pacman_object : public GameAbstractObject, public Pacman
 {
 
 private:
-    int lastShoot;
+    long long lastShoot;
 
 protected:
 
@@ -16,8 +16,8 @@ public:
     pacman_object(int type);
     void setPosition(const QPointF& _pos);
     void set_angle(qreal angle);
-    int lastShooting() const;
-    int& lastShooting();
+    long long lastShooting() const;
+    long long& lastShooting();
 };
 
 #endif // PACMAN_OBJECT_H
